Fix int32_t min issues

+++ src/TMC5160.cpp
@@ -217,7 +217,7 @@ void TMC5160::setTargetPosition(float position)
 
 void TMC5160::setMaxSpeed(float speed)
 {
-	writeRegister(TMC5160_Reg::VMAX, min(0x7FFFFF, speedFromHz(fabs(speed)))); // VMAX : 23 bits
+	writeRegister(TMC5160_Reg::VMAX, min(int32_t(0x7FFFFF), speedFromHz(fabs(speed)))); // VMAX : 23 bits
 
 	if (_currentRampMode == VELOCITY_MODE)
 	{
@@ -227,27 +227,27 @@ void TMC5160::setMaxSpeed(float speed)
 
 void TMC5160::setRampSpeeds(float startSpeed, float stopSpeed, float transitionSpeed)
 {
-	writeRegister(TMC5160_Reg::VSTART, min(0x3FFFF, speedFromHz(fabs(startSpeed)))); // VSTART : 18 bits
-	writeRegister(TMC5160_Reg::VSTOP, min(0x3FFFF, speedFromHz(fabs(stopSpeed)))); // VSTOP : 18 bits
-	writeRegister(TMC5160_Reg::V_1, min(0xFFFFF, speedFromHz(fabs(transitionSpeed)))); // V1 : 20 bits
+	writeRegister(TMC5160_Reg::VSTART, min(int32_t(0x3FFFF), speedFromHz(fabs(startSpeed)))); // VSTART : 18 bits
+	writeRegister(TMC5160_Reg::VSTOP, min(int32_t(0x3FFFF), speedFromHz(fabs(stopSpeed)))); // VSTOP : 18 bits
+	writeRegister(TMC5160_Reg::V_1, min(int32_t(0xFFFFF), speedFromHz(fabs(transitionSpeed)))); // V1 : 20 bits
 }
 
 void TMC5160::setAcceleration(float maxAccel)
 {
-	writeRegister(TMC5160_Reg::AMAX, min(0xFFFF, accelFromHz(fabs(maxAccel)))); // AMAX, DMAX: 16 bits
-	writeRegister(TMC5160_Reg::DMAX, min(0xFFFF, accelFromHz(fabs(maxAccel))));
+	writeRegister(TMC5160_Reg::AMAX, min(int32_t(0xFFFF), accelFromHz(fabs(maxAccel)))); // AMAX, DMAX: 16 bits
+	writeRegister(TMC5160_Reg::DMAX, min(int32_t(0xFFFF), accelFromHz(fabs(maxAccel))));
 }
 
 void TMC5160::setAccelerations(float maxAccel, float maxDecel, float startAccel, float finalDecel)
 {
-	writeRegister(TMC5160_Reg::AMAX, min(0xFFFF, accelFromHz(fabs(maxAccel)))); // AMAX, DMAX, A1, D1 : 16 bits
-	writeRegister(TMC5160_Reg::DMAX, min(0xFFFF, accelFromHz(fabs(maxDecel))));
-	writeRegister(TMC5160_Reg::A_1, min(0xFFFF, accelFromHz(fabs(startAccel))));
-	writeRegister(TMC5160_Reg::D_1, min(0xFFFF, accelFromHz(fabs(finalDecel))));
+	writeRegister(TMC5160_Reg::AMAX, min(int32_t(0xFFFF), accelFromHz(fabs(maxAccel)))); // AMAX, DMAX, A1, D1 : 16 bits
+	writeRegister(TMC5160_Reg::DMAX, min(int32_t(0xFFFF), accelFromHz(fabs(maxDecel))));
+	writeRegister(TMC5160_Reg::A_1, min(int32_t(0xFFFF), accelFromHz(fabs(startAccel))));
+	writeRegister(TMC5160_Reg::D_1, min(int32_t(0xFFFF), accelFromHz(fabs(finalDecel))));
 }
 
 /**
- * 
+ *
  * @see Datasheet rev 1.15, section 6.3.2.2 "RAMP_STAT - Ramp & Reference Switch Status Register".
  * @return true if the target position has been reached, false otherwise.
  */
@@ -259,7 +259,7 @@ bool TMC5160::isTargetPositionReached(void)
 }
 
 /**
- * 
+ *
  * @see Datasheet rev 1.15, section 6.3.2.2 "RAMP_STAT - Ramp & Reference Switch Status Register".
  * @return true if the target velocity has been reached, false otherwise.
  */
@@ -290,7 +290,7 @@ void TMC5160::enable()
 	writeRegister(TMC5160_Reg::CHOPCONF, _chopConf.value);
 }
 
-TMC5160::DriverStatus TMC5160::getDriverStatus() 
+TMC5160::DriverStatus TMC5160::getDriverStatus()
 {
 	TMC5160_Reg::GSTAT_Register gstat = {0};
 	gstat.value = readRegister(TMC5160_Reg::GSTAT);
@@ -338,9 +338,9 @@ const char* TMC5160::getDriverStatusDescription(DriverStatus st)
 
 void TMC5160::setModeChangeSpeeds(float pwmThrs, float coolThrs, float highThrs)
 {
-	writeRegister(TMC5160_Reg::TPWMTHRS, min(0xFFFFF, thrsSpeedToTstep(pwmThrs))); // 20 bits
-	writeRegister(TMC5160_Reg::TCOOLTHRS, min(0xFFFFF, thrsSpeedToTstep(coolThrs)));
-	writeRegister(TMC5160_Reg::THIGH, min(0xFFFFF, thrsSpeedToTstep(highThrs)));
+	writeRegister(TMC5160_Reg::TPWMTHRS, min(int32_t(0xFFFFF), thrsSpeedToTstep(pwmThrs))); // 20 bits
+	writeRegister(TMC5160_Reg::TCOOLTHRS, min(int32_t(0xFFFFF), thrsSpeedToTstep(coolThrs)));
+	writeRegister(TMC5160_Reg::THIGH, min(int32_t(0xFFFFF), thrsSpeedToTstep(highThrs)));
 }
 
 bool TMC5160::setEncoderResolution(int32_t motorSteps, int32_t encResolution, bool inverted)
